Non-security bugfix; streaming validation added to (meta_)ratchet
Summary: Every operation is supposed to have a sanity check built in where it
asserts that, if the `more` flag is set, the current operation is the same as
the previous. This was implemented in every function except for ratchet.
Obviously, this check should be implemented for ratchet too, so I implemented it
and added a regression test.